XAKATON-main/innopolis_emergency_bot/requirements.txt
aiogram>=3.26,<4.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
google-genai>=1.0,<2.0
psycopg[binary]>=3.2
groq>=0.11.0
